backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "emails>=0.6",
    "fastapi>=0.121.3",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.22",
    "ruff>=0.14.6",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]

[tool.ruff]
lint.select = ["E", "W", "F", "I", "B", "UP", "PL", "RUF", "COM", "SLF", "TID"]
lint.ignore = [
    "E712",
    "B904",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF010",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR2004",
]
line-length = 100
exclude = [
    ".venv",
    "venv",
    "__pycache__",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"
